expense-related classes moved to expenses subpackage
