+++ src/test/java/com/centrilli/pages/expenses/ExpenseReportsPage.java
@@ -1,4 +1,4 @@
-package com.centrilli.pages;
+package com.centrilli.pages.expenses;
 
 import com.centrilli.utilities.Driver;
 import org.openqa.selenium.WebElement;
